refactor(web-core): move env var validation to utils
